Relocate custom serializers to a dedicated data serialization package
This includes:
- Moving `HtmlDecodedStringSerializer` and `LogEntrySerializer` from the `utils` package to a new `data.serialization` package.
- Updating package declarations and imports across multiple API model classes, including `Timers`, `Events`, `Services`, `Movies`, `DeviceInfo`, and `SignalInfo`, to reflect the new location.
