+++ app/src/main/java/io/github/deprec8/enigmadroid/data/serialization/HtmlDecodedStringSerializer.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.utils
+package io.github.deprec8.enigmadroid.data.serialization
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
