+++ app/src/main/java/io/github/deprec8/enigmadroid/model/api/DeviceInfo.kt
@@ -20,7 +20,7 @@
 package io.github.deprec8.enigmadroid.model.api
 
 import androidx.compose.runtime.Immutable
-import io.github.deprec8.enigmadroid.utils.HtmlDecodedStringSerializer
+import io.github.deprec8.enigmadroid.data.serialization.HtmlDecodedStringSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
