+++ app/src/main/java/io/github/deprec8/enigmadroid/model/api/Timers.kt
@@ -20,8 +20,8 @@
 package io.github.deprec8.enigmadroid.model.api
 
 import androidx.compose.runtime.Immutable
-import io.github.deprec8.enigmadroid.utils.HtmlDecodedStringSerializer
-import io.github.deprec8.enigmadroid.utils.LogEntrySerializer
+import io.github.deprec8.enigmadroid.data.serialization.HtmlDecodedStringSerializer
+import io.github.deprec8.enigmadroid.data.serialization.LogEntrySerializer
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.SerialName
